+++ roseau/load_flow/models/buses.py
@@ -292,7 +292,7 @@ class Bus(AbstractTerminal[CyBus]):
             ground:
                 If a ground is given, the phases will also be connected to the ground.
         """
-        from roseau.load_flow import PowerLoad
+        from roseau.load_flow import CurrentLoad, PowerLoad
 
         for phase in phases:
             if phase not in self.phases:
@@ -314,9 +314,10 @@ class Bus(AbstractTerminal[CyBus]):
             logger.error(msg)
             raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
         for element in self._connected_elements:
-            if isinstance(element, PowerLoad):
+            if isinstance(element, (PowerLoad, CurrentLoad)):
+                load_type = "power" if isinstance(element, PowerLoad) else "current"
                 msg = (
-                    f"A power load {element.id!r} is already connected on bus {self.id!r}. "
+                    f"A {load_type} load {element.id!r} is already connected on bus {self.id!r}. "
                     f"It makes the short-circuit calculation impossible."
                 )
                 logger.error(msg)