+++ roseau/load_flow_single/models/buses.py
@@ -212,17 +212,18 @@ class Bus(AbstractTerminal[CyBus]):
 
     @property
     def short_circuit(self) -> bool:
-        """Return the list of short-circuits of this bus."""
+        """Whether there is a short circuit on the bus or not"""
         return self._short_circuit
 
     def add_short_circuit(self) -> None:
         """Add a short-circuit by connecting all the phases together with a ground."""
-        from roseau.load_flow_single import PowerLoad
+        from roseau.load_flow_single import CurrentLoad, PowerLoad
 
         for element in self._connected_elements:
-            if isinstance(element, PowerLoad):
+            if isinstance(element, (PowerLoad, CurrentLoad)):
+                load_type = "power" if isinstance(element, PowerLoad) else "current"
                 msg = (
-                    f"A power load {element.id!r} is already connected on bus {self.id!r}. "
+                    f"A {load_type} load {element.id!r} is already connected on bus {self.id!r}. "
                     f"It makes the short-circuit calculation impossible."
                 )
                 logger.error(msg)