+++ roseau/load_flow_single/models/tests/test_buses.py
@@ -5,6 +5,7 @@ import pytest
 from roseau.load_flow import Q_, RoseauLoadFlowException, RoseauLoadFlowExceptionCode
 from roseau.load_flow_single import (
     Bus,
+    CurrentLoad,
     ElectricalNetwork,
     Line,
     LineParameters,
@@ -40,6 +41,15 @@ def test_short_circuit():
     bus = Bus("bus")
     assert not bus.short_circuit
     _ = PowerLoad(id="load", bus=bus, power=10)
+    with pytest.raises(RoseauLoadFlowException) as e:
+        bus.add_short_circuit()
+    assert "is already connected on bus" in e.value.msg
+    assert e.value.code == RoseauLoadFlowExceptionCode.BAD_SHORT_CIRCUIT
+
+    # Cannot short-circuit a bus with a current load
+    bus = Bus("bus")
+    assert not bus.short_circuit
+    _ = CurrentLoad(id="load", bus=bus, current=10)
     with pytest.raises(RoseauLoadFlowException) as e:
         bus.add_short_circuit()
     assert "is already connected on bus" in e.value.msg